separate express app and nodejs server into two files (app.ts and server.ts); create upgradeHandler.ts file for handling upgrade events; change run script in package.json so that it starts server.ts file; update import paths
