+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "build": "tsc && tsc-alias",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon --watch src --exec ts-node --require tsconfig-paths/register src/index.ts",
+    "dev": "nodemon --watch src --exec ts-node --require tsconfig-paths/register src/server.ts",
     "start": "node dist/index.js"
   },
   "keywords": [],
